Refactor: Client 전달 및 send 수정
