+++ src/components/chat/ChatSection.tsx
@@ -19,21 +19,19 @@ function ChatSection() {
 
   const { menuModalState } = useMenuModalState();
   const [showShoutBubble, setShowShoutBubble] = useState(true);
-  const [stompClient, setStompClient] = useState<Client | null>(null);
+  const [stompClient, setStompClient] = useState<Client>();
 
   useEffect(() => {
     const socket = configureStompClient(ROOMNUM);
 
-    socket.onConnect = (frame: frame) => {
-      console.log("connected", frame);
+    socket.onConnect = () => {
       setStompClient(socket);
       socket.subscribe(
         `/sub/chat/${ROOMNUM}`,
         (frame) => {
           try {
-            console.log("구독요청");
-            // const receivedMessage = JSON.parse(frame.body);
-            console.log(frame);
+            const receivedMessage = JSON.parse(frame.body);
+            console.log(receivedMessage);
           } catch (error) {
             console.error("stomp 구독에 오류가 발생했습니다:", error);
           }
@@ -45,32 +43,8 @@ function ChatSection() {
       );
     };
 
-    socket.onStompError = (frame) => {
-      console.log("Broker reported error: " + frame.headers["message"]);
-      console.log("Additional details: " + frame.body);
-    };
-
     socket.activate();
   }, []);
-
-  const sendMessage = (msg: any) => {
-    if (stompClient && stompClient.connected) {
-      const messageDetails = JSON.stringify({
-        gameId: `${ROOMNUM}`,
-        type: "NORMAL",
-        content: msg,
-      });
-
-      stompClient.publish({
-        destination: "/pub/chat/message",
-        body: messageDetails,
-        headers: {
-          gameId: `${ROOMNUM}`,
-          Authorization: `${localStorage.getItem("authToken")}`,
-        },
-      });
-    }
-  };
 
   return (
     <div className="chat">
@@ -81,7 +55,7 @@ function ChatSection() {
             {menuModalState.isOpen && <MenuModal />}
             <VoteModal />
           </div>
-          <ChatInput sendMessage={sendMessage} />
+          <ChatInput stompClient={stompClient} />
         </div>
         <ChatLog />
       </div>
